app completed but for the live stock data and lederboard

+++ db/migrate/20210226122750_create_holdings.rb
@@ -3,7 +3,7 @@ class CreateHoldings < ActiveRecord::Migration[6.1]
     create_table :holdings do |t|
       t.integer :owner_users_ID, null: false
       t.string :stock_code, null: false
-      t.integer :quantity, null: false
+      t.float :quantity, null: false
       t.integer :asking, null: false
 
       t.timestamps
